Cleans up closure conversion and adds more tests
Note: During free variable calculation, another pass is performed
to keep track of the number of arguments for f.

Note: If a function has more arguments than the number of registers,
or if adding the closure function code pointer results in too many
arguments, it will not spill those arguments to the closure yet. That
means you cannot have more than (# of registers - 1) arguments for any
function.